--- PackageSelector.axaml
@@ -1,37 +1,0 @@
-<Window xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d" d:DesignWidth="500" d:DesignHeight="600"
-        x:Class="AndroidDebloater.PackageSelector"
-        xmlns:local="using:AndroidDebloater"
-        Width="500" Height="600"
-        Title="Select Packages to remove">
-    
-        <Grid Margin="10">
-                <Grid.RowDefinitions>
-                        <RowDefinition Height="*"/>
-                        <RowDefinition Height="*"/>
-                        <RowDefinition Height="auto"/>
-                </Grid.RowDefinitions>
-                
-                <ScrollViewer Grid.Row="0" MaxHeight="500" HorizontalScrollBarVisibility="Hidden" VerticalScrollBarVisibility="Auto" Margin="0 10 0 10">
-                        <ItemsControl Name="PackageList">
-                                <ItemsControl.ItemTemplate>
-                                        <DataTemplate DataType="{x:Type local:AndroidPackage}">
-                                                <CheckBox Content="{Binding Text}"
-                                                          IsChecked="{Binding IsChecked}"
-                                                          Margin="0,5,0,0"/>
-                                        </DataTemplate>
-                                </ItemsControl.ItemTemplate>
-                        </ItemsControl>
-                </ScrollViewer>
-                <TextBox x:Name="LogBox" Grid.Row="1" Background="Black" Foreground="White" IsReadOnly="True"/>
-                <Button Grid.Row="2" Name="ShowSelectedButton"
-                        Content="Uninstall selected"
-                        Margin="0,10,0,0"
-                        Click="RemoveSelected"/>
-                
-        </Grid>
-        
-</Window>
